requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
amqp==5.3.1
anyio==4.11.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpickle==3.1.2
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cron-descriptor==1.4.5
cycler==0.12.1
Django==4.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-graphql-jwt==0.4.0
django-timezone-field==7.1
filelock==3.20.0
fiona==1.10.1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.10.0
geopandas==1.1.1
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
GraphQL-core-next==1.1.1
graphql-relay==3.2.0
graphql-ws-next==1.0.0
gunicorn==23.0.0
h11==0.16.0
holidays==0.73
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.3
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numba==0.62.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.1
prophet==1.1.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
redis==6.2.0
regex==2025.11.3
requests==2.32.3
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sqlparse==0.5.3
stanio==0.5.1
statsmodels==0.14.4
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
xgboost==3.0.2
yarl==1.20.0
